Update README.md
In this lab you will be writing a dynamic memory allocator for C programs; that is, your own version of the malloc, free, realloc functions. You are encouraged to explore the design space creatively and implement an allocator that is correct, efficient, and fast.

Materials

You can start the assignment by downloading prj4-malloc.tar from cyber campus to a directory in which you plan to do your work. Then give the command tar -xvf prj4-malloc.tar. This will cause a number of files to be unpacked into the directory.

Requirements: The only file you will be modifying and turning in ismm.c, which contains your solution.

Test Driver Program: The mdriver.c program is a driver program that allows you to locally evaluate the performance of your solution in the same way that TA will evaluate your final submission. Use the command make to generate the driver code and run it with the command ./mdriver.

Project Description

Your dynamic memory allocator will consist of the following functions, which are declared in mm.h and defined in mm.c: